refactor dependencies
